catalogos/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.5.4
pytest==8.2.2
reportlab==4.0.7
requests==2.31.0
gunicorn==20.1.0
psycopg2-binary==2.9.9